FIX refactor auth y restringiendo lo subido a github
