+++ storage/framework/.gitignore
@@ -7,3 +7,4 @@ routes.php
 routes.scanned.php
 schedule-*
 services.json
+/cache
