--- storage/framework/cache/data/60/25/60258c940f591d805fdfefc02a56944855011790
@@ -1,1 +1,0 @@
-1711568855i:1;
